+++ tests/ApiTest.php
@@ -2,16 +2,26 @@
 
 class ApiTest extends TestCase
 {
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->artisan('grafite:starter')
+            ->expectsQuestion('Are you sure you want to overwrite any files of the same name?', true)
+            ->expectsQuestion('Would you like to run the migration?', false)
+            ->assertExitCode(0);
+    }
+
     public function testApiCommand()
     {
-        $status = $this->app['Illuminate\Contracts\Console\Kernel']->handle(
-            $input = new \Symfony\Component\Console\Input\ArrayInput([
-                'command' => 'grafite:api',
-                '--no-interaction' => true
-            ]),
-            $output = new \Symfony\Component\Console\Output\BufferedOutput
-        );
+        $this->artisan('grafite:api')
+            ->expectsQuestion('Are you sure you want to overwrite any files of the same name?', true)
+            ->assertExitCode(0);
+    }
 
-        $this->assertContains('php artisan grafite:starter', $output->fetch());
+    public function testFilesExist()
+    {
+        $this->assertTrue(file_exists(base_path('app/Http/Controllers/Api/AuthController.php')));
+        $this->assertTrue(file_exists(base_path('app/Http/Controllers/Api/UserController.php')));
     }
 }