+++ tests/NotificationsTest.php
@@ -2,16 +2,27 @@
 
 class NotificationsTest extends TestCase
 {
-    public function testNotificationsCommandWithoutStarter()
+    public function setUp()
     {
-        $status = $this->app['Illuminate\Contracts\Console\Kernel']->handle(
-            $input = new \Symfony\Component\Console\Input\ArrayInput([
-                'command' => 'grafite:notifications',
-                '--no-interaction' => true
-            ]),
-            $output = new \Symfony\Component\Console\Output\BufferedOutput
-        );
+        parent::setUp();
 
-        $this->assertContains('php artisan grafite:starter', $output->fetch());
+        $this->artisan('grafite:starter')
+            ->expectsQuestion('Are you sure you want to overwrite any files of the same name?', true)
+            ->expectsQuestion('Would you like to run the migration?', false)
+            ->assertExitCode(0);
+    }
+
+    public function testBootstrapCommand()
+    {
+        $this->artisan('grafite:notifications')
+            ->expectsQuestion('Are you sure you want to overwrite any files of the same name?', true)
+            ->assertExitCode(0);
+    }
+
+    public function testFilesExist()
+    {
+        $this->assertTrue(file_exists(base_path('app/Http/Controllers/Admin/NotificationController.php')));
+        $this->assertTrue(file_exists(base_path('app/Http/Controllers/User/NotificationController.php')));
+        $this->assertTrue(file_exists(base_path('app/Services/NotificationService.php')));
     }
 }